refactor(client): extract lsp.client to a top-level client package
The connected-LSP-client state (ClientCapabilitiesHolder, LanguageClientHolder,
WorkDoneProgressHelper) is a dependency-free leaf used widely — by providers,
context, the lsp services, codelenses and more. Sitting under lsp.client, it
folded into the lsp slice, so every consumer's edge counted as an edge into lsp
and closed the lsp<->providers / context<->lsp loops, fusing 21 packages into a
single cycle.

Moving it to a top-level client package (a pure leaf) detaches those edges: lsp,
providers and almost every feature package drop out of the cycle. The package
dependency graph collapses from one 21-package cycle to a layered DAG with just
two small cycles left — the 7-package domain core
(configuration/context/diagnostics/types/references/infrastructure/formatting)
and codelenses<->commands. The defunct unconstrained LspClient layer is dropped
and client is added to the acyclic-domain set.
