+++ src/test/java/com/github/_1c_syntax/bsl/languageserver/infrastructure/LanguageClientAwareAppenderTest.java
@@ -23,7 +23,7 @@ package com.github._1c_syntax.bsl.languageserver.infrastructure;
 
 import ch.qos.logback.classic.Level;
 import ch.qos.logback.classic.spi.ILoggingEvent;
-import com.github._1c_syntax.bsl.languageserver.lsp.client.LanguageClientHolder;
+import com.github._1c_syntax.bsl.languageserver.client.LanguageClientHolder;
 import org.eclipse.lsp4j.MessageParams;
 import org.eclipse.lsp4j.MessageType;
 import org.eclipse.lsp4j.services.LanguageClient;
